+++ tests/Unit/Cache/PsrSimpleCacheAdapterTest.php
@@ -71,7 +71,7 @@ final class PsrSimpleCacheAdapterTest extends TestCase
     public function test_custom_key_factory(): void
     {
         $cache = new InMemorySimpleCache();
-        $keyFactory = fn (string $regex) => 'custom_'.md5($regex);
+        $keyFactory = static fn (string $regex) => 'custom_'.md5($regex);
         $adapter = new PsrSimpleCacheAdapter($cache, prefix: 'test_', keyFactory: $keyFactory);
 
         $key = $adapter->generateKey('/test/');